FOGL-3473 Solve issue with FIONREAD return type

+++ C/services/storage/include/stream_handler.h
@@ -62,7 +62,7 @@ class StreamHandler {
 									m_pool;
 					};
 					void		setNonBlocking(int fd);
-					size_t		available(int fd);
+					int		available(int fd);
 					void		queueInsert(StorageApi *api, unsigned int nReadings, bool commit);
 					void		dump(int n);
 					enum { Closed, Listen, AwaitingToken, Connected }
